requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# Gravitational field/field.py: 1
Pillow == 8.2.0

# Gravitational field/cl.py: 2
# Gravitational field/field.py: 2
numpy == 1.20.4

# Gravitational field/cl.py: 1
pyopencl == 2021.1.6
